fix(db): CockroachDB PERCENTILE_CONT decimal→float8 전면 수정
PERCENTILE_CONT에 decimal 타입 전달 시 CockroachDB에서
unknown signature: percentile_cont_impl(decimal, decimal) 오류 발생.
ORDER BY 컬럼에 ::float8 캐스트 추가 (6개 파일).
a_value.py except 블록에 db.rollback() 추가하여 트랜잭션 오염 방지.

+++ bid-system/backend/app/api/v1/admin.py
@@ -1404,7 +1404,7 @@ def ml_calibration(
         SELECT
             AVG(ABS(srate_error))  AS mae,
             STDDEV(srate_error)    AS std,
-            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY srate_error) AS median_bias
+            PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY srate_error::float8) AS median_bias
         FROM bid_journal
         WHERE srate_error IS NOT NULL
     """)).fetchone()
